telco_churn_rate/deployment/backend/requirements.txt
waitress
pandas
flask
tensorflow-cpu
scikit-learn == 1.0.2